Isolated sign up logic into a util file
